+++ .gitignore
@@ -42,3 +42,4 @@ yarn-error.log*
 # OS Files
 .DS_Store
 
+.vscode/launch.json
