+++ flask-backend/chatbot_routes.py
@@ -1,5 +1,5 @@
 import os
-from typing import Annotated, TypedDict, Literal
+from typing import Annotated, TypedDict, Literal, List
 from dotenv import load_dotenv
 from langchain import hub
 from flask import request, jsonify,Blueprint
@@ -15,22 +15,75 @@ from langchain_community.agent_toolkits.sql.toolkit import SQLDatabaseToolkit
 from langchain_community.tools.sql_database.tool import QuerySQLDatabaseTool
 from llm_classes import State, Assistant, is_statistical_question, start_node
 from langgraph.types import Command
+from langchain.schema import AIMessage
+from sqlalchemy import create_engine, inspect
+from datetime import datetime
 
 
-chatbot_bp = Blueprint("chatbot_bp", __name__)
+
+chatbot_bp = Blueprint("chatbot_bp1", __name__)
 
 load_dotenv()
 
+def log_with_time(msg):
+    now = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    print(f"{msg} [Time: {now}]")
+
 # # ----- LLM and Database Setup ----- #
 
-db_username = os.getenv('DB_USERNAME') # admin
-db_password = os.getenv('DB_PASSWORD') # asabasketball
-db_name = os.getenv('DB_NAME') # game_stats
-db_host = os.getenv('DB_HOST') # ucd-basketball.cduqug2e0o83.us-east-2.rds.amazonaws.comprompt_template = hub.pull("langchain-ai/sql-agent-system-prompt")
+#db_username = os.getenv('DB_USERNAME') # admin
+#db_password = os.getenv('DB_PASSWORD') # asabasketball
+#db_name = os.getenv('DB_NAME') # game_stats
+#db_host = os.getenv('DB_HOST') # ucd-basketball.cduqug2e0o83.us-east-2.rds.amazonaws.comprompt_template = hub.pull("langchain-ai/sql-agent-system-prompt")
 
+
+llm = ChatOpenAI(model="gpt-3.5-turbo", temperature=0.3)
+#db_uri = f"mysql+mysqlconnector://{db_username}:{db_password}@{db_host}/{db_name}"
+#db = SQLDatabase.from_uri(db_uri)
+#table_info = db.get_table_info()
+
+
+#THIS IS THE NEW SQLITE DATABASE CONNECTION CODE
+base_dir = os.path.abspath(os.path.dirname(__file__))
+db_path = os.path.join(base_dir, "ucd-basketball.db")
+db_uri = f"sqlite:///{db_path}"
+db = SQLDatabase.from_uri(db_uri)
+
+
+#THIS IS JUST TO TEST TO SEE IF SQLITE DATABASE IS CONNECTED AND WORKING
+engine = create_engine(db_uri)
+inspector = inspect(engine)
+table_names = inspector.get_table_names()
+print(table_names)
+
+if db: 
+    try: 
+        print("DB connection successful from chatbot routes 2")
+            # Try running a basic test query to make sure the DB is responsive
+        table_names = db.get_usable_table_names()
+        print("✅ DB connection successful. Tables found:", table_names)
+    except Exception as e:
+        print("❌ Failed to connect to DB:", str(e))
+
+table_info = db.get_table_info()
+# print(table_info)
+
+MAX_RETRIES = 3
+
+class State(TypedDict):
+    question: str
+    query: str
+    result: str
+    answer: str
+    retry: bool  # flag indicating to retry if there's an error
+    retry_count: int   # counter for retry attempts
+    error: str 
+
+
+# ---- Query Agent (State Graph) Setup ----- #
 
 prompt_template = hub.pull("langchain-ai/sql-agent-system-prompt")
-prompt_template.messages[0].pretty_print()
+# system_message.messages[0].pretty_print()
 """
 You are an agent designed to interact with a SQL database.
 Given an input question, create a syntactically correct {dialect} query to run, then look at the results of the query and return the answer.
@@ -48,165 +101,61 @@ Do NOT skip this step.
 Then you should query the schema of the most relevant tables.
 """
 
-
-llm = ChatOpenAI(model="gpt-3.5-turbo", temperature=0.3)
-#db_uri = f"mysql+mysqlconnector://{db_username}:{db_password}@{db_host}/{db_name}"
-#db = SQLDatabase.from_uri(db_uri)
-
-base_dir = os.path.abspath(os.path.dirname(__file__))
-db_path = os.path.join(base_dir, "ucd-basketball.db")
-db_uri = f"sqlite:///{db_path}"
-
-# Use it to initialize SQLDatabase
-db = SQLDatabase.from_uri(db_uri)
-if db: 
-    print("db connection successful from chatbot routes")
+toolkit = SQLDatabaseToolkit(db=db, llm=llm)
+tools = toolkit.get_tools()
 
 
-#app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///ucd-basketball.db'
+class QueryQuestionsOutput(TypedDict):
+    questions: List[str]
 
-table_info = db.get_table_info()
-print(table_info)
-
-
-query_prompt_template = hub.pull("langchain-ai/sql-query-system-prompt")
-
-MAX_RETRIES = 3
-
-class State(TypedDict):
-    question: str
-    query: str
-    result: str
-    answer: str
-    retry: bool  # flag indicating to retry if there's an error
-    retry_count: int   # counter for retry attempts
-    error: str 
-
-
-class QueryOutput(TypedDict):
-    """Generated SQL query."""
-
-    query: Annotated[str, ..., "Syntactically valid SQL query."]
-
-
-# ---- Query Agent (State Graph) Setup ----- #
-
-def determine_relevant_stats(state: State) -> str:
-    """Determine relevant stats to query based on the user question."""
-    prompt = (
-        "You are a UC Davis basketball analyst and scout. "
-        "Given the following user question, determine the relevant stats that should be queried to provide an insightful answer. "
-        "If the user question is asking for a specific stat, output that stat in a way that can be easily translated into a SQL query. "
-        "If the question is ambiguous, list a few relevant stats that would help provide a detailed response.\n\n"
-        f"Question: {state['question']}"
-    )
-    response = llm.invoke(prompt)
-    return response.content
-
-def write_query(state: State):
-    """Generate SQL query to fetch information, informed by relevant stats."""
-    relevant_stats = determine_relevant_stats(state)
-    # You might log or print the relevant stats for debugging:
-    print("Relevant stats determined:", relevant_stats)
+def query_decision_agent(state: dict) -> Command[Literal["generate_answer", END]]:
+    """
+    Given the user's question, determine which specific database queries would help answer it.
+    If the question is directly about a specific stat, output a single query.
+    If it's general, output multiple query questions.
     
-    if state["retry"] and state.get("retry_count", 0) < MAX_RETRIES:
-        state["retry_count"] += 1
-        prompt_text = (
-            "The previous SQL query resulted in the error: "
-            f"{state.get('error', 'Unknown error')}\n"
-            "Please generate an alternative SQL query that avoids this error. "
-            f"Question: {state['question']} "
-            f"and consider these relevant stats: {relevant_stats}"
-        )
-    else:
-        table_info = db.get_table_info()
-        prompt_text = query_prompt_template.invoke({
-            "dialect": db.dialect,
-            "top_k": 10,
-            "table_info": table_info,
-            "input": f"{state['question']} Relevant stats: {relevant_stats}",
-        })
-    structured_llm = llm.with_structured_output(QueryOutput)
-    result = structured_llm.invoke(prompt_text)
-    return {"query": result["query"]}
-
-def execute_query(state: State):
-    """Execute SQL query.
-    If an error is detected (e.g. missing column), set a retry flag.
+    Expected output format:
+      { "questions": ["<SQL-like question 1>", "<SQL-like question 2>", ...] }
     """
-    # breakpoint()  
 
-    execute_query_tool = QuerySQLDatabaseTool(db=db)
-    result = execute_query_tool.invoke(state["query"])
-
-    if isinstance(result, str) and ("unknown column" in result.lower() or "error" in result.lower()):
-        state["retry"] = True
-        state["error"] = result
-        state["retry_count"] = state.get("retry_count", 0) + 1
-        return {"result": "", "retry": state["retry"], "retry_count": state["retry_count"], "error": state["error"]}
-    else:
-        state["retry"] = False
-        return {"result": result, "retry": state["retry"], "retry_count": state["retry_count"], "error": state["error"]}
-
-def generate_answer(state: State):
-    """Answer question using retrieved information as context."""
-    # breakpoint()
+    #breakpoint()
 
     prompt = (
-        "You are a UC Davis basketball analyst and scout."
-        "Given the following user question, corresponding SQL query, "
-        "and SQL result, provide a nuanced answer to the question. "
-        "Only mention the stats in you response, do not mention the SQL query."
-        "\n\n"
-        f'Question: {state["question"]}\n'
-        f'SQL Query: {state["query"]}\n'
-        f'SQL Result: {state["result"]}'
+        "You are a UC Davis Basketball analyst and scout. Your task is to determine which database queries will provide the most useful insights based on the user’s input. \n\n"
+        "- If the question is already a direct request for a single piece of data (for example: \"Who is the leading scorer on UC Davis?\"), just rephrase the question (if necessary).  \n"
+        "- Keep in mind, per game stats are generally more useful than season totals."
+        "- If the question is more general or exploratory (for example: \"Give me a scouting report on UC Riverside\"), break it down into multiple detailed query questions that would cover relevant trends, player statistics, and performance metrics.  \n"
+        "- The output should always be a JSON object with a list of questions. \n\n"
+        "You have access to the following table information:\n"
+        "{table_info}\n\n"
+        "Based on this, for the given user question below, decide whether to return it as-is or to break it down into multiple queries, then provide your output in the specified JSON format.\n\n"
+        "User Question:\n"
+        f"{state.get('question')}\n\n"
+        "Output format:\n"
+        "{ \"questions\": [\"<query question 1>\", \"<query question 2>\", ...] }"
     )
+
+    structured_llm = llm.with_structured_output(QueryQuestionsOutput)
+    result = structured_llm.invoke(prompt)
+    log_with_time(f"[QueryDecisionAgent] Agent generated query questions: {result}")
+    return result
+
+def generate_answer(state: dict) -> Command[Literal["end"]]:
+    """
+    Generate the final answer to the user's question using the database query results.
+    """
+
+    prompt = (
+        "You are a UC Davis Basketball analyst and scout. "
+        "Given the following question and the results of the database queries, generate a detailed and actionable insight that answers the user's question.\n\n"
+        f"Question: {state.get('question')}\n"
+        f"Database result: {state.get('relevant_stats')}\n\n"
+        "Output your answer in a clear and concise manner, and only use stats that are relevant to the question."
+    )
+
     response = llm.invoke(prompt)
-
-    if isinstance(response, str) and ("error" in response.lower() or "unknown column" in response.lower() or "modify" in response.lower()):
-        state["retry"] = True
-        state["error"] = response
-        state["retry_count"] = state.get("retry_count", 0) + 1
-        return {"answer": "", "retry": state["retry"], "retry_count": state["retry_count"], "error": state["error"]}
-    state["answer"] = response.content
-    return {"answer": response.content}
-
-def execute_checker(state: State):
-    """Determine the next node based on state conditions"""
-    if state.get("retry", False) and state.get("retry_count", 0) < MAX_RETRIES:
-        return "write_query"
-    return "generate_answer"
-
-def answer_checker(state: State):
-    """Determine the next node based on state conditions"""
-    if state.get("retry", False) and state.get("retry_count", 0) < MAX_RETRIES:
-        return "write_query"
-    return END
-
-def needs_db_query(state: State) -> bool:
-    """
-    Function to determine if a database query is needed based on keywords in user input
-    """
-    keywords = ["scorer", "points", "stat", "game", "compare", "leader"]
-    question = state["question"].lower()
-    if any(keyword in question for keyword in keywords):
-        return "write_query"
-    return "generate_answer"
-
-graph_builder = StateGraph(State)
-graph_builder.add_node("write_query", write_query)
-graph_builder.add_node("execute_query", execute_query)
-graph_builder.add_node("generate_answer", generate_answer)
-
-graph_builder.add_edge(START, "write_query")
-graph_builder.add_edge("write_query", "execute_query")
-graph_builder.add_conditional_edges("execute_query", execute_checker)
-graph_builder.add_conditional_edges("generate_answer", answer_checker)
-query_graph = graph_builder.compile()
-
-# ----- Supervisor Agent ----- #
-
+    log_with_time(f"[GenerateAnswer] LLM generated answer: {response.content}")
+    return response.content
 
 def direct_answer(question: str) -> str:
     """
@@ -214,7 +163,10 @@ def direct_answer(question: str) -> str:
     """
     prompt = f"As a UC Davis Basketball analyst, answer the following question: {question}"
     response = llm.invoke(prompt)
+    log_with_time(f"[DirectAnswer] LLM generated direct answer: {response.content}")
     return response.content
+
+# ----- Supervisor Agent ----- #
 
 def supervisor(state: dict) -> Command[Literal["direct_answer", "db_query", END]]:
     """
@@ -244,31 +196,83 @@ def supervisor(state: dict) -> Command[Literal["direct_answer", "db_query", END]
         next_agent = END
     else:
         next_agent = "direct_answer"
+    log_with_time(f"[Supervisor] Supervisor agent decided next step: {next_agent}")
     return Command(goto=next_agent)
 
 def overarching_supervisor(state: dict) -> dict:
-    
-    cmd = supervisor(state)
-    if cmd.goto == "direct_answer":
-        # Call the direct answer function
-        answer = direct_answer(state["question"])
-        return {"response": answer, "path": "direct"}
-    elif cmd.goto == "db_query":
-        # Run the query graph and extract the answer
-        current_state = state.copy()
-        steps = []
-        for step in query_graph.stream(state):
-            steps.append(step)
-            print(step)
-            if isinstance(step, dict):
-                current_state.update(step)
-        answer = current_state.get("generate_answer", {}).get("answer", "")
-        print(current_state)
-        if not answer:
-            answer = "I'm sorry, I couldn't generate a response from the database."
-        return {"response": answer, "path": "db_query", "debug_info": {"steps": steps, "final_state": current_state}}
-    else:  # if END or other commands
-        return {"response": "Goodbye!", "path": "end"}
+    # State
+    try:
+        cmd = supervisor(state)
+        if cmd.goto == "direct_answer":
+            # Call the direct answer function
+            answer = direct_answer(state["question"])
+            log_with_time(f"[OverarchingSupervisor] Direct answer generated: {answer}")
+            return {"response": answer, "path": "direct"}
+        elif cmd.goto == "db_query":
+            # Run the query graph and extract the answer
+            query_spec_output = query_decision_agent(state)
+            query_questions = query_spec_output.get("questions", [])
+            log_with_time(f"[OverarchingSupervisor] Query questions to execute: {query_questions}")
+            system_message = prompt_template.format(dialect="MySQL", top_k=5)
+            agent_executor = create_react_agent(llm, tools, prompt=system_message)
+
+            relevant_stats = []
+            query_errors = []
+
+            for question in query_questions:
+                try:
+                    result = agent_executor.invoke({
+                        "messages": [
+                            {
+                                "role": "user",
+                                "content": question
+                            }
+                        ]
+                    })
+                    ai_message = next(
+                        (msg.content for msg in reversed(result['messages']) if isinstance(msg, AIMessage)), 
+                        None
+                    )
+                    log_with_time(f"[OverarchingSupervisor] Agent executed query: {question} Response: {ai_message}")
+                    if ai_message:
+                        relevant_stats.append(ai_message)
+                    
+                except Exception as e:
+                    query_errors.append(f"Error querying '{question}': {str(e)}")
+
+            state["relevant_stats"] = relevant_stats
+            state["query_errors"] = query_errors
+
+            if relevant_stats:
+                answer = generate_answer(state)
+                log_with_time(f"[OverarchingSupervisor] Final answer generated from DB queries: {answer}")
+                return {
+                        "response": answer,
+                        "path": "db_query",
+                        "status": "success",
+                        "metadata": {
+                            "queries_executed": len(query_questions),
+                            "successful_queries": len(relevant_stats),
+                            "failed_queries": len(query_errors)
+                        }
+                    }
+            else:
+                log_with_time(f"[OverarchingSupervisor] All queries failed. Errors: {query_errors}")
+                return {
+                        "response": "I encountered issues while querying the database. Please try rephrasing your question.",
+                        "path": "db_query",
+                        "status": "error",
+                        "errors": query_errors
+                    }
+    except Exception as e:
+        log_with_time(f"[OverarchingSupervisor] Unexpected error: {str(e)}")
+        return {
+            "response": "I apologize, but I encountered an unexpected error while processing your request.",
+            "path": "error",
+            "status": "error",
+            "error": str(e)
+        }
+
 
 
 @chatbot_bp.route('/chat', methods=['POST', 'OPTIONS'])
@@ -277,19 +281,16 @@ def chat():
         return jsonify({"status": "ok"}), 200
         
     try:
-        breakpoint()
+        #breakpoint()
 
         data = request.json
         user_message = data.get("message")
 
         state: State = { # initial state
             "question": user_message,
-            "query": "",
+            "relevant_stats": "",
             "result": "",
             "answer": "",
-            "retry": False,
-            "retry_count": 0,
-            "error": ""
         }
 
         #call overarching agent
@@ -302,4 +303,3 @@ def chat():
             "error": "An error occurred while processing your request",
             "details": str(e)
         }), 500
-
